Let's keep the "extending Rust class" strategy for now.
