+++ libraries/matrix/impl/src/test/kotlin/io/element/android/libraries/matrix/impl/room/member/RoomMemberListFetcherTest.kt
@@ -15,7 +15,7 @@ import io.element.android.libraries.matrix.api.room.roomMembers
 import io.element.android.libraries.matrix.impl.room.member.RoomMemberListFetcher.Source.CACHE
 import io.element.android.libraries.matrix.impl.room.member.RoomMemberListFetcher.Source.CACHE_AND_SERVER
 import io.element.android.libraries.matrix.impl.room.member.RoomMemberListFetcher.Source.SERVER
-import io.element.android.libraries.matrix.impl.sdk.FakeRoomInterface
+import io.element.android.libraries.matrix.impl.sdk.FakeRoom
 import io.element.android.libraries.matrix.impl.sdk.FakeRoomMembersIterator
 import io.element.android.libraries.matrix.test.A_USER_ID
 import io.element.android.libraries.matrix.test.A_USER_ID_2
@@ -31,7 +31,7 @@ import uniffi.matrix_sdk.RoomMemberRole
 class RoomMemberListFetcherTest {
     @Test
     fun `fetchRoomMembers with CACHE source - emits cached members, if any`() = runTest {
-        val room = FakeRoomInterface(getMembersNoSync = {
+        val room = FakeRoom(getMembersNoSync = {
             FakeRoomMembersIterator(
                 listOf(
                     fakeRustRoomMember(A_USER_ID),
@@ -62,7 +62,7 @@ class RoomMemberListFetcherTest {
 
     @Test
     fun `fetchRoomMembers with CACHE source - emits empty list, if no members exist`() = runTest {
-        val room = FakeRoomInterface(getMembersNoSync = {
+        val room = FakeRoom(getMembersNoSync = {
             FakeRoomMembersIterator(emptyList())
         })
 
@@ -77,7 +77,7 @@ class RoomMemberListFetcherTest {
 
     @Test
     fun `fetchRoomMembers with CACHE source - emits Error on error found`() = runTest {
-        val room = FakeRoomInterface(getMembersNoSync = {
+        val room = FakeRoom(getMembersNoSync = {
             error("Some unexpected issue")
         })
 
@@ -92,7 +92,7 @@ class RoomMemberListFetcherTest {
 
     @Test
     fun `fetchRoomMembers with CACHE source - emits all items at once`() = runTest {
-        val room = FakeRoomInterface(getMembersNoSync = {
+        val room = FakeRoom(getMembersNoSync = {
             FakeRoomMembersIterator(
                 listOf(
                     fakeRustRoomMember(A_USER_ID),
@@ -119,7 +119,7 @@ class RoomMemberListFetcherTest {
 
     @Test
     fun `fetchRoomMembers with SERVER source - emits only new members, if any`() = runTest {
-        val room = FakeRoomInterface(getMembers = {
+        val room = FakeRoom(getMembers = {
             FakeRoomMembersIterator(
                 listOf(
                     fakeRustRoomMember(A_USER_ID),
@@ -145,7 +145,7 @@ class RoomMemberListFetcherTest {
 
     @Test
     fun `fetchRoomMembers with SERVER source - on error it emits an Error item`() = runTest {
-        val room = FakeRoomInterface(getMembers = { error("An unexpected error") })
+        val room = FakeRoom(getMembers = { error("An unexpected error") })
 
         val fetcher = RoomMemberListFetcher(room, Dispatchers.Default)
         fetcher.membersFlow.test {
@@ -159,7 +159,7 @@ class RoomMemberListFetcherTest {
 
     @Test
     fun `fetchRoomMembers with CACHE_AND_SERVER source - returns cached items first, then new ones`() = runTest {
-        val room = FakeRoomInterface(
+        val room = FakeRoom(
             getMembersNoSync = {
                 FakeRoomMembersIterator(listOf(fakeRustRoomMember(A_USER_ID_4)))
             },
